refactor: separate setup and helper functions in their own files
